migrated functionality from jaxampler to gwkokab

+++ gwkokab/_src/models/eccmodel.py
@@ -19,18 +19,15 @@ from typing import Optional
 from jax.random import truncated_normal
 from jaxtyping import Array
 
-from ..typing import Numeric
 from ..utils.misc import get_key
 from .abstracteccentricitymodel import AbstractEccentricityModel
 
 
 class EccentricityModel(AbstractEccentricityModel):
-    def __init__(self, sigma_ecc: Numeric, name: Optional[str] = None) -> None:
-        # super().__init__(loc=0.0, scale=sigma_ecc, low=0.0, high=1.0, name=name)
+    def __init__(self, sigma_ecc: float, name: Optional[str] = None) -> None:
         self._sigma_ecc = sigma_ecc
 
     def samples(self, num_of_samples: int) -> Array:
-        # return super().rvs(shape=(num_of_samples,), key=None)
         return truncated_normal(key=get_key(), lower=0.0, upper=1.0, shape=(num_of_samples,)) * self._sigma_ecc
 
     def __repr__(self) -> str:
